Follow-on to 7bdccade98e76df3a7830bc45bc42321a77c709a to exclude directories for now.

+++ modules/gallery/helpers/gallery_event.php
@@ -54,6 +54,11 @@ class gallery_event_Core {
               continue;
             }
 
+            // Ignore directories for now, but we should really address them in the long term.
+            if (is_dir("$dir/$file")) {
+              continue;
+            }
+
             if (filemtime("$dir/$file") <= $threshold) {
               unlink("$dir/$file");
               break;
